feat: Improve rust integer type definition
Look at minimum and maximum for integers to identify
i8/u8/i16/u16/i32/u32


Changes are triggered by https://review.opendev.org/960272

+++ openstack_types/src/block_storage/v3/backup/response/list_detailed.rs
@@ -103,7 +103,7 @@ pub struct BackupResponse {
 
     /// The size of the volume, in gibibytes (GiB).
     #[structable(wide)]
-    pub size: i64,
+    pub size: i32,
 
     /// The UUID of the source volume snapshot.
     #[serde(default)]
