+++ openstack_types/src/block_storage/v3/volume/response/set.rs
@@ -172,7 +172,7 @@ pub struct VolumeResponse {
 
     /// The size of the volume, in gibibytes (GiB).
     #[structable()]
-    pub size: i64,
+    pub size: i32,
 
     /// To create a volume from an existing snapshot, specify the UUID of the
     /// volume snapshot. The volume is created in same availability zone and
